requirements.txt
streamlit~=1.45.0
pandas~=2.2.3
numpy~=2.2.5
scikit-learn~=1.6.1
matplotlib~=3.10.1
seaborn~=0.13.2
python-docx>=1.0.0
openpyxl>=3.1.2
joblib>=1.3.2

scipy~=1.15.2